Renames list parts None fixture for better reuse
